+++ CHANGES.rst
@@ -3,6 +3,7 @@
 - Added ``owners`` search param.
 - Added ``q`` search param.
 - Added group ``~banned``.
+- Wrapped ``User:PATCH``.
 
 3.0.3
 -----